Oppdatert samtlige sider
La til tre bilder i wwwroot/images. Gitt bilde til bakgrunn på landingPage og UserPage. Oppdatert UserPageModel. I UserPage.cshtml, er det oppdatert struktur med tabell i en frame og en sort-and-filter funksjon i den andre, oppdatert css i henhold til dette.
