+++ Gruppe6_Kartverket.Mvc/Models/Database/CaseLocation.cs
@@ -8,6 +8,6 @@
         public string County { get; set; }
 
         // Navigation property to list of Cases
-        public virtual ICollection<CaseRecord> CaseRecordss { get; set; }
+        public virtual ICollection<CaseRecord> CaseRecords { get; set; }
     }
 }
