+++ Gruppe6_Kartverket.Mvc/Models/Database/CaseRecord.cs
@@ -3,8 +3,11 @@
     public class CaseRecord
     {
         public int CaseId { get; set; }
+        
         public DateTime CaseDate { get; set; }
+        
         public string CaseDescription { get; set; }
+        
         public string CaseStatus { get; set; }
         public int LocationId { get; set; }
         public int UserId { get; set; }
