--- Gruppe6_Kartverket.Mvc/Models/ErrorViewModel.cs
@@ -1,8 +1,0 @@
-namespace Gruppe6_Kartverket.Mvc.Models;
-
-public class ErrorViewModel
-{
-    public string? RequestId { get; set; }
-
-    public bool ShowRequestId => !string.IsNullOrEmpty(RequestId);
-}
