--- Gruppe6_Kartverket.Mvc/Models/SettingsModel.cs
@@ -1,8 +1,0 @@
-namespace Gruppe6_Kartverket.Mvc.Models;
-
-public class SettingsModel
-{
-    public string Username { get; set; }
-    public string Email { get; set; }
-    public string Password { get; set; }
-}
