+++ Gruppe6_Kartverket.Mvc/Models/ViewModels/ErrorViewModel.cs
@@ -1,0 +1,8 @@
+namespace Gruppe6_Kartverket.Mvc.Models;
+
+public class ErrorViewModel
+{
+    public string? RequestId { get; set; }
+
+    public bool ShowRequestId => !string.IsNullOrEmpty(RequestId);
+}
