+++ Gruppe6_Kartverket.Mvc/Models/ViewModels/NotificationModel.cs
@@ -1,0 +1,7 @@
+namespace Gruppe6_Kartverket.Mvc.Models;
+
+public class Notification
+{
+    public string Message { get; set; }
+    public DateTime Date { get; set; }
+}
