+++ Gruppe6_Kartverket.Mvc/Models/ViewModels/SettingsModel.cs
@@ -1,0 +1,8 @@
+namespace Gruppe6_Kartverket.Mvc.Models;
+
+public class SettingsModel
+{
+    public string Username { get; set; }
+    public string Email { get; set; }
+    public string Password { get; set; }
+}
