+++ Gruppe6_Kartverket.Mvc/Views/CaseWorker/CaseWorkerPage.cshtml
@@ -155,7 +155,7 @@
             </div>
         </div>
     </div>
-</div>
+
 
 
 @section Scripts {
