+++ Gruppe6_Kartverket.Mvc/Views/ContactUs/ContactUs.cshtml
@@ -1,5 +1,3 @@
-@model Gruppe6_Kartverket.Mvc.Models.ContactUsModel 
-
 <div class="site-wrapper">
     <h1>Kontakt oss</h1>
