+++ Gruppe6_Kartverket.Mvc/wwwroot/css/landingPage.css
@@ -2,7 +2,9 @@
 .landing-page {
     text-align: center;
     padding: 2rem;
-    background-color: #D0ECD6; /* Light green background color */
+    background-image: url("../Images/kartverket-bakgrunssbilder/norway.jpg"); /* Path to the background image */
+    background-size: cover; /* Ensure the image covers the entire background */
+    background-position: center; /* Center the background image */
 }
 
 /* Wrapper for centering content on the page */
